ZDF uses semicolon in video description
ZDF uses semicolon in video description. The description is added with
other parameters into an URL. When parsing and spliting the URL in
python2 the semicolon is used as a separator which leads to a wrong data
in the resulting dictionary and video cannot be found by video player.

Semicolon is replaced by "." instead and everything works again in
python2